Fix SuperApp Payment. Change amount parameter from Int to Double (#2)

+++ Android/MiniAppCaller/payments/src/main/java/com/genexus/superapps/bankx/payments/services/PaymentsService.kt
@@ -12,7 +12,7 @@ object PaymentsService {
     private const val SDT_PAYMENT_INFORMATION_ITEM_AFFINITY = "affinity"
     private const val SDT_PAYMENT_INFORMATION_ITEM_TYPE = "type"
 
-    fun pay(amount: Int): String {
+    fun pay(amount: Double): String {
         return amount.toString() + "-" + UUID.randomUUID().toString()
     }
